moving-average/requirements.txt
tabulate
numpy
pandas
matplotlib
python-binance
pandas_datareader